test_requirements.txt
django-nose
coverage
django-coverage
django-jasmine
ipdb
flake8
fabric
factory_boy
mock
selenium
coveralls
django-crispy-forms
